requirements.txt
asgiref==3.8.1
Django==5.0.4
pillow==10.3.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
django-import-export==4.0.7
